backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Caching==2.0.2
Flask-Limiter==3.5.0
Flask-Uploads==0.2.1
SQLAlchemy==2.0.21
PyMySQL==1.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
celery==5.3.4
redis==5.0.1
flasgger==0.9.7
python-jose==3.3.0
requests==2.31.0
pillow==10.0.1
python-magic==0.4.27
boto3==1.28.36
stripe==6.5.0 